organized the profiles and their inspectors a bit more. Fixed a few misc things
