+++ Assets/MixedRealityToolkit-SDK/Profiles/SpatialAwareness/DefaultWindowsMixedRealitySpatialMeshObserverProfile.asset
