+++ Assets/MixedRealityToolkit/_Core/Inspectors/Profiles/SpatialAwareness.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 9d3f310959bc2f546a8a642a01a8d183
+guid: 45a5c7370af8a6e409970d7b40f3d344
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
