+++ Assets/MixedRealityToolkit/_Core/Inspectors/Profiles/SpatialAwareness/BaseMixedRealitySpatialObserverProfileInspector.cs
